// src/app/emrks/model/doctor.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Doctor {
    id?: string;
    name?: string;
    startDate?: string;
    specialty?: string;
    officeName?: string;
    dssn?: string;
}